GET from API working
